Fix StreamChannel bug
Instead of accumulating a buffer at the writer end, it is now
accumulated at the reader end. All traffic is routed through
the "fetcher" asynchronous generator. This fixes the bug that
blocked the TestStreamChannel.test_read_all() unit test.

The current implementation passes all the tests in the current
test suite.