Implement Channel using a single Future field (#31)
This commit replaces the implementation based on
itertools.tee() with a very simple implementation
that does not generate or iterate over futures at
all. Instead, it creates an ordinary Python class
with a Future field and an asend() method that
sets, replaces, and awaits that field in order.

This works because even though it looks like an
asynchronous generator, it does not share its
limitation of not being able to pick up a value
from the second call to asend before yielding a
return value to the awaiting first call to
asend. The asend method is "reentrant", so to
speak.