Factor out .abc and .futures
Without changing the dynamic API or the tests, this factors out the
underlying classes in .abc and .futures and removes those submodules
altogether.

Statically, things are much different now: if c is a Channel (but not
a StreamChannel), then the client() and server() methods no longer
provide access to separate asynchronous generators. They have become
dummy methods, returning self to maintain duck type compatibility
with StreamChannel.

In a future revision, if the Writer and Reader classes are fused into
the StreamChannel class in a similar manner, the client() and server()
methods could be removed from the API.