StreamChannel now single object too, API revision (#32)
This commit applies the same philosophy from the previous
commit to the StreamChannel class. It is still a simplex
channel, but it is now large duck-type compatible with
an asynchronous generator for the receiving end and with
a reverse asynchronouse generator for the sending end.

The client() and server() methods have been removed from
the Channel and StreamChannel APIs and the tests have
been modified accordingly.