Rename stream methods (#38)
This commit fixes issue #38.

The following methods have been renamed:
 * Reader.read_once() -> receive()
 * Reader.read_items() -> read()
 * Writer.write_items() -> write()

The current names were the result of an attempt to follow a convention
in the Python ecosystem of using method names like readframes() or
writelines() for stream methods that accept or return a specific type of
object instead of strings or bytes.

However, the channel stream methods do not actually return a specific
type of object. The whole point of these methods is that they return a
generic type of object. If you see a call to readframes in a piece of
code that tells you something about what is being read. If you see a
call to read_items() you don't know anything. The only thing it tells
you is that you are not getting a single bytes or string object back.
But that is considerably less useful information.

Furthermore, note that the names "read_items" and "write_items" already
were a compromise between the "readfoo" and "writebar" convention on the
one hand, and the general naming convention in Python to use underscores
on the other hand. Because "items" starts with a vowel, "writeitems" is
much more difficult to read than, for example, "writelines".

Finally, note that the "readfoo" naming convention predates type
annotation. One of the drawbacks of type hinting is that you get really
long function signatures. For existing libraries, using stub files means
that the typed signatures don't pollute your code, and also that you can
use a different norm for line length in .pyi files.

However, for new libraries where you use type hinting in the .py code
itself, the degree to which type hinting makes your code look cluttered
and Java-esque depends largely on how much your annotations are
repeating your function and parameter names.

For example, if you are designing a MessageWriter, then without type
annotations, a method name like write_message(m) would be fine. And if
the function body would be long and contain several local variables,
then you might even want to call it write_message(msg) or
write_message(message), depending on whether you like abbreviations.

However, with type annotations, the result might be as follows:

    class MessageWriter:
        def write_message(message: Message) -> None:
            ...

Instead, type annotations make the following declaration unambiguous:

    class MessageWriter:
        def write(m: Message) -> None:
            ...

So it seems to me that when we apply type annotation in the design stage
of an API, as opposed to afterwards, we should consider trading those
annotations for shorter method and parameter names.